Generalize nouns-to-gf.py to handle compound nouns
Also, the 11-character limit on the word length was removed.

make-dict.sh now uses a file (not checked in) that was
generated by Morfessor 2.0 by
  - training it on all the WordNet nouns and then
  - using the obtained model to segment the same nouns.

As a result most of the compound nouns in DictEst now
get the correct representation. There are a few mistakes as well,
e.g. 'kokkaja' -> mkN "kokk" (mkN "aja" "aja" "ajat" ...)
(which do not necessarily result in incorrect forms though).

We should evaluate if we'd get a better dictionary by using
Filosoft's compound splitting.